Time: 24 ms (90.09%), Space: 63.2 MB (60.03%) - LeetHub

+++ 1335-maximum-candies-allocated-to-k-children/1335-maximum-candies-allocated-to-k-children.java
@@ -1,16 +1,16 @@
 class Solution {
     public int maximumCandies(int[] candies, long k) {
 
-        long start = 1;
-        long end = Integer.MIN_VALUE;
+        int start = 1;
+        int end = Integer.MIN_VALUE;
         for(int c:candies) {
             end = Math.max(end, c);
         }
 
-        long result = 0;
+        int result = 0;
 
         while(start <= end) {
-            long mid = start + (end - start)/2;
+            int mid = start + (end - start)/2;
 
             if(checkCandies(candies, mid, k)) {
                 result = mid;
@@ -20,10 +20,10 @@ class Solution {
                 end = mid - 1;
             }
         }
-        return (int)result;
+        return result;
     }
 
-     public static boolean checkCandies(int[] candies, long mid, long k) {
+     public static boolean checkCandies(int[] candies, int mid, long k) {
         long eatingCandies = 0;
         for(int candy : candies) {
             eatingCandies = eatingCandies  + (candy/mid);
